requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
httpx==0.27.2
trafilatura==1.12.0
beautifulsoup4==4.12.3
jinja2==3.1.4
aiofiles==24.1.0
lxml==5.3.0
pypdf==5.1.0
rapidfuzz>=3.9
scikit-learn>=1.5
requests>=2.32
